requirements.txt
amqp==5.2.0
billiard==4.2.0
blinker==1.8.2
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
MarkupSafe==2.1.5
marshmallow==3.21.3
mysqlclient==2.2.4
packaging==24.0
prompt_toolkit==3.0.46
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.5
scheduler==0.8.5
secure-smtplib==0.1.1
six==1.16.0
SQLAlchemy==2.0.30
typeguard==4.3.0
typing_extensions==4.12.1
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
